Move `test_st_pyplot` and `test_st_pyplot_clear_figure` out of streamlit_test.py and into pyplot_test.py
